Fixes for readr-1.2.0

+++ tests/testthat/test_col_types.R
@@ -2,7 +2,7 @@ test_that("col_types ODS", {
     x <- read_ods('../testdata/col_types.ods', col_types = NA)
     expect_equal(class(x[,2]), "character")
     x <- read_ods('../testdata/col_types.ods')
-    expect_equal(class(x[,2]), "integer")
+    expect_equal(class(x[,2]), "numeric")
 })
 
 ### test for issue #41
